Move SelectionList to base_picking_dialog to eliminate duplication across alignment dialogs
